Updates and fixes
Now the x scale is based on the prior. If the sample data is left as
NULL then it plots only the prior distribution. If sample data is
inputted (mean, SE), then updates plot with lines and legend for
likelihood and posterior, and returns posterior summary. If null value
is specified it adds null dots to the plot and calculates BFs.